Move CognitoStack to dedicated package
